requirements.txt
pyperclip
requests
packaging
streamlit
pywebview